selenium: No more config files
Change the `selenium` command to use a `selenium_port` file in the same
directory as the script (and add it to gitignore). We no longer need
`_setup`. Make use of Bash functions and local variables.

If you want to specify a custom SSH command, you can do it with, e.g.:

ssh="mysshcmd -f" ./selenium

I thought it was still worthwhile to save the port in a file because it
needs to be set up on the remote machine, and it would be annoying to
have to change that every time.

This also adds some instructions for how to set up the port on the
remote machine so that unit tests know which port to use, and it
switches to using `$!` to get the PID of the last _background_ process.