feat(selenium): add a selenium grid config file
This configures the local grid a bit more to make it easier to use locally.

It will immediately reject any new session requests if there are no matching
browsers.

It will only add drivers we have explicitly declared and configured instead of
automatically adding all drivers on the system. On top of that, it will use
the chromedriver binary we downloaded via the `get_selenium.sh` script.