drop selenium-config.toml template
A corresponding pull request will be opened in the iFixit/ifixit repo to drop
the browser version from the new session request. This will allow the Distributor
to choose a browser just matching the browser name and platform.

As such, we don't need to template the selenium-config.toml file anymore.